Remove inner classes and static methods
Remove inner classes from MainActivity.java and create separate classes. Communicate between Fragment and Activities through Interface instead of accessing MainActivity directly from Fragments or Adapters.
