+++ .gitignore
@@ -39,6 +39,9 @@ captures/
 .idea/gradle.xml
 .idea/dictionaries
 .idea/libraries
+.idea/compiler.xml
+.idea/misc.xml
+.idea/modules.xml
 
 # Keystore files
 *.jks
