+++ app/src/main/java/manparvesh/ideatrackerplus/IdeaActivityHost.java
@@ -1,0 +1,11 @@
+package manparvesh.ideatrackerplus;
+
+/**
+ * IdeaActivityHost  Listener implemented by the MainActivity
+ */
+
+public interface IdeaActivityHost {
+    boolean isSearchEnabled();
+
+    void openProjectDialog();
+}
