+++ app/src/main/java/manparvesh/ideatrackerplus/ItemAdapter.java
@@ -1,6 +1,6 @@
 package manparvesh.ideatrackerplus;
 
-/**
+/*
  * Copyright 2014 Magnus Woxblom
  * <p/>
  * Licensed under the Apache License, Version 2.0 (the "License");
